+++ packages/f-button/src/components/tests/Button.test.js
@@ -1,0 +1,10 @@
+import { shallowMount } from '@vue/test-utils';
+import VueButton from '../Button.vue';
+
+describe('Button', () => {
+    it('should be defined', () => {
+        const propsData = {};
+        const wrapper = shallowMount(VueButton, { propsData });
+        expect(wrapper.exists()).toBe(true);
+    });
+});