+++ packages/f-button/test-utils/component-objects/f-button.component.js
@@ -1,0 +1,5 @@
+const buttonComponent = () => $('[data-test-id="button-component"]');
+
+exports.waitForButtonComponent = () => buttonComponent().waitForExist();
+
+exports.isButtonComponentDisplayed = () => buttonComponent().isDisplayed();