+++ packages/f-content-cards/vue.config.js
@@ -12,5 +12,8 @@ module.exports = {
                 importer: magicImporter(),
                 data: '@import "@/assets/scss/common.scss";'
             });
+    },
+    pluginOptions: {
+        lintStyleOnBuild: true
     }
 };