+++ packages/f-metadata/src/BrazeDispatcher.js
@@ -45,7 +45,8 @@ function contentCardsHandler (postCardsAppboy) {
 
     const {
         cards,
-        rawCards
+        rawCards,
+        groups
     } = new ContentCards(postCardsAppboy, { enabledCardTypes: this.dispatcherOptions.enabledCardTypes })
         .removeDuplicateContentCards()
         .filterCards()
@@ -54,6 +55,7 @@ function contentCardsHandler (postCardsAppboy) {
         .output();
 
     this.cardsCallbacks.forEach(callback => callback(cards));
+    this.groupedCardsCallback.forEach(callback => callback(groups));
 
     this.rawCards = rawCards;
 
@@ -95,6 +97,7 @@ class BrazeDispatcher {
         this.dispatcherOptions = null;
 
         this.cardsCallbacks = [];
+        this.groupedCardsCallback = [];
         this.inAppMessageClickEventCallbacks = [];
         this.inAppMessagesCallbacks = [];
 
@@ -147,6 +150,7 @@ class BrazeDispatcher {
         if (callbacks.interceptInAppMessageClickEvents) this.inAppMessageClickEventCallbacks.push(callbacks.interceptInAppMessageClickEvents);
         if (callbacks.interceptInAppMessages) this.inAppMessagesCallbacks.push(callbacks.interceptInAppMessages);
         if (callbacks.handleContentCards) this.cardsCallbacks.push(callbacks.handleContentCards);
+        if (callbacks.handleContentCardsGrouped) this.groupedCardsCallback.push(callbacks.handleContentCardsGrouped);
 
         // Note that appBoyPromise will not be set if this is the first time running this method -
         // this is a guard against initialise() being called more than once, and attempting to