+++ packages/f-registration/src/components/Registration.vue
@@ -1,15 +1,15 @@
 <template>
-    <div>
-        <bag-celebrate-icon :class="$style['bag-icon']" />
+    <div
+        :class="$style['c-registration']">
         <card
             :data-theme-registration="theme"
             :card-heading="copy.labels.createAccountTitle"
             is-rounded
-            has-outline
             is-page-content-wrapper
             card-heading-position="center"
             data-test-id="registration-component"
             :class="$style['c-card-padding']">
+            <bag-celebrate-icon :class="$style['bag-icon']" />
             <p
                 v-if="showLoginLink"
                 :class="$style['c-loginLink']"
@@ -261,6 +261,11 @@ export default {
             type: String,
             required: true
         },
+        createAccountTimeout: {
+            type: Number,
+            required: false,
+            default: 1000
+        },
         showLoginLink: {
             type: Boolean,
             default: true
@@ -408,24 +413,32 @@ export default {
                     registrationSource: 'Native',
                     marketingPreferences: []
                 };
-                await RegistrationServiceApi.createAccount(this.createAccountUrl, this.tenant, registrationData);
+                await RegistrationServiceApi.createAccount(this.createAccountUrl, this.tenant, registrationData, this.createAccountTimeout);
                 this.$emit(EventNames.CreateAccountSuccess);
             } catch (error) {
                 let thrownErrors = error;
                 if (error && error.response && error.response.data && error.response.data.errors) {
                     thrownErrors = error.response.data.errors;
                 }
+                let shouldEmitCreateAccountFailure = true;
 
                 if (Array.isArray(thrownErrors)) {
                     if (thrownErrors.some(thrownError => thrownError.errorCode === '409')) {
                         this.shouldShowEmailAlreadyExistsError = true;
+                    } else if (thrownErrors.some(thrownError => thrownError.errorCode === 'FailedUserAuthentication')) {
+                        this.$emit(EventNames.LoginBlocked);
+
+                        shouldEmitCreateAccountFailure = false;
                     } else {
                         this.genericErrorMessage = thrownErrors[0].description || 'Something went wrong, please try again later';
                     }
                 } else {
                     this.genericErrorMessage = error;
                 }
-                this.$emit(EventNames.CreateAccountFailure, thrownErrors);
+
+                if (shouldEmitCreateAccountFailure) {
+                    this.$emit(EventNames.CreateAccountFailure, thrownErrors);
+                }
             } finally {
                 this.shouldDisableCreateAccountButton = false;
             }
@@ -443,6 +456,9 @@ export default {
 <style lang="scss" module>
 
 // Form styling
+.c-registration {
+    margin-top: 100px;
+}
 
 .o-form {
     @include font-size(body-l);
@@ -466,31 +482,36 @@ export default {
     margin-top: spacing(x2);
 }
 
-.c-loginLink {
+.c-loginLink,
+.c-legal-hyperlinks {
     text-align: center;
     a {
         color: $blue;
         text-decoration: none;
-    }
-}
-
-.c-legal-hyperlinks {
-    a {
-        color: $blue;
-        text-decoration: none;
+        font-weight: $font-weight-bold;
     }
 }
 
 .c-card-padding {
-    padding-top: 30px;
+    padding-top: spacing(x5);
+    padding-bottom: spacing(x6);
+
+    @include media('<mid') {
+        padding-bottom: spacing(x4);
+    }
 }
 
 .bag-icon {
     width: 97px;
     height: 78px;
-    margin: auto;
-    display: block;
-    position: relative;
-    bottom: -30px;
+    position: absolute;
+    top: 56px;
+    left: 50%;
+    transform: translate(-35%);
+    @include media('<mid') {
+        width: 92px;
+        height: 74px;
+    }
 }
+
 </style>