+++ packages/f-registration/src/components/tests/f-registration.integration.test.js
@@ -25,6 +25,7 @@ describe('Registration API service', () => {
 
     afterEach(() => {
         wrapper.destroy();
+        jest.clearAllMocks();
     });
 
     it('responds with 201 when request is made with valid details', async () => {
@@ -73,5 +74,34 @@ describe('Registration API service', () => {
         // Assert
         expect(wrapper.emitted(EventNames.CreateAccountFailure).length).toBe(1);
         expect(wrapper.vm.shouldShowEmailAlreadyExistsError).toBe(true);
+    });
+
+    it('responds with 403 when login blocked by ravelin or recaptcha', async () => {
+        // Arrange
+        axiosMock.onPost(propsData.createAccountUrl, CONSUMERS_REQUEST_DATA).reply(403, {
+            faultId: '25bbe062-c53d-4fbc-9d6c-3df6127b94fd',
+            traceId: 'H3TKh4QSJUSwVBCBqEtkKw',
+            errors: [
+                {
+                    description: 'Failed user authentication.',
+                    errorCode: 'FailedUserAuthentication'
+                }
+            ]
+        });
+
+        // Act
+        wrapper.find('[data-test-id="input-first-name"]').setValue(CONSUMERS_REQUEST_DATA.firstName);
+        wrapper.find('[data-test-id="input-last-name"]').setValue(CONSUMERS_REQUEST_DATA.lastName);
+        wrapper.find('[data-test-id="input-email"]').setValue(CONSUMERS_REQUEST_DATA.emailAddress);
+        wrapper.find('[data-test-id="input-password"]').setValue(CONSUMERS_REQUEST_DATA.password);
+        // TODO: Add marketing preferences checkbox when it exists
+
+        // Act
+        await wrapper.vm.onFormSubmit();
+        await flushPromises();
+
+        // Assert
+        expect(wrapper.emitted(EventNames.LoginBlocked).length).toBe(1);
+        expect(wrapper.emitted(EventNames.CreateAccountFailure)).toBeUndefined();
     });
 });