+++ packages/f-form-field/CHANGELOG.md
@@ -5,7 +5,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 v0.8.0
 ------------------------------
-*October 21, 2020*
+*October 23, 2020*
 
 ### Added 
 - data-test IDs to FormField Component